Moved dynamic series functions to line series.
